Goncharenko

+++ app/Http/Controllers/Users/UserController.php
@@ -4,23 +4,18 @@ namespace App\Http\Controllers\Users;
 
 use App\Http\Controllers\Controller;
 use App\Http\Filters\Users\UserFilter;
-use App\Http\Requests\Tasks\StoreTaskFormRequest;
-use App\Http\Requests\Tasks\UpdateTaskFormRequest;
 use App\Http\Requests\Users\StoreUserFormRequest;
 use App\Http\Requests\Users\UpdateUserFormRequest;
 use App\Http\Requests\Users\UserFilterRequest;
-use App\Models\Documents\Document;
 use App\Models\UserRoles\UserRole;
-use App\Services\Tasks\UploadService;
-use Illuminate\Http\Request;
 
-use App\Models\Tasks\Task;
-use App\Models\Tasks\TaskPriority;
 use App\Models\Roles\Role;
 use App\Models\User;
-use Illuminate\Support\Benchmark;
+
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Log;
 use Symfony\Polyfill\Uuid\Uuid;
 
 
@@ -39,7 +34,14 @@ class UserController extends Controller
      */
     public function index(UserFilterRequest $request)
     {
-        //$this->authorize('viewAny', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+                'request' => $request->all(),
+
+            ]);
+
+        $this->authorize('viewAny', User::class);
 
         $data = $request->validated();
 
@@ -61,7 +63,12 @@ class UserController extends Controller
      */
     public function create()
     {
-        //$this->authorize('create', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+            ]);
+
+        $this->authorize('create', User::class);
 
         return view('users.create', [
             'superiors' => User::all(),
@@ -77,7 +84,14 @@ class UserController extends Controller
      */
     public function store(StoreUserFormRequest $request)
     {
-        //$this->authorize('create', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+                'request' => $request->all(),
+
+            ]);
+
+        $this->authorize('create', User::class);
 
         if ($request->isMethod('post'))
         {
@@ -120,8 +134,7 @@ class UserController extends Controller
 
             } catch (\Exception $e) {
                 DB::rollBack();
-                dd($e); // TODO сделать вывод в журнол ошибок, чтобы сайт не крашился
-            }
+                Log::error($e);            }
         }
 
         return redirect()->route('users.create')->with('error', 'Не удалось создать нового сотрудника.');
@@ -135,7 +148,13 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        //$this->authorize('view', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+                //'user_request_data' => $user
+            ]);
+
+        $this->authorize('view', User::class);
 
         return view('users.show', [
             'user' => $user,
@@ -151,7 +170,12 @@ class UserController extends Controller
      */
     public function edit(User $user)
     {
-        //$this->authorize('update', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+            ]);
+
+        $this->authorize('update', User::class);
 
         return view('users.edit', [
             'user' => $user,
@@ -170,7 +194,14 @@ class UserController extends Controller
      */
     public function update(UpdateUserFormRequest $request, User $user)
     {
-        //$this->authorize('update', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+                'request' => $request->all(),
+
+            ]);
+
+        $this->authorize('update', User::class);
 
         if($request->isMethod('patch')){
 
@@ -209,8 +240,7 @@ class UserController extends Controller
 
             } catch (\Exception $e) {
                 DB::rollBack();
-                dd($e); // TODO, вывод ошибки в журнал, чтобы сайт не крашился
-            }
+                Log::error($e);            }
         }
         return redirect()->route('users.edit', $user->id)->with('error', 'Не удалось сохранить новые данные.');
     }
@@ -223,7 +253,12 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        //$this->authorize('delete', User::class);
+        Log::info(get_class($this) . ', method: ' . __FUNCTION__,
+            [
+                'user' => Auth::user()->name,
+            ]);
+
+        $this->authorize('delete', User::class);
 
         $user->delete();
         return redirect()->route('users.index');
